@@ -1,14 +1,14 @@
 {
-  "name": "app-insights-traces-client",
+  "name": "application-insights-client",
   "version": "0.1.0",
   "dependencies": {
-    "react": "^16.3.1",
-    "react-dom": "^16.3.1",
+    "react": "^16.5.2",
+    "react-dom": "^16.5.2",
     "react-redux": "^5.0.7",
-    "react-scripts": "1.1.4",
-    "redux": "^3.7.2",
-    "redux-observable": "^0.18.0",
-    "rxjs": "^5.5.10"
+    "react-scripts": "2.0.4",
+    "redux": "^4.0.1",
+    "redux-observable": "^1.0.0",
+    "rxjs": "^6.3.3"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -29,6 +29,12 @@
   },
   "homepage": "https://joanjane.github.io/application-insights-client",
   "devDependencies": {
-    "gh-pages": "^1.1.0"
-  }
+    "gh-pages": "^2.0.1"
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
