@@ -1,10 +1,15 @@
-import { Observable } from 'rxjs/Observable';
+import { of } from 'rxjs';
+import { switchMap } from 'rxjs/operators';
+import { ofType } from 'redux-observable';
 import { SET_CREDENTIALS, availableAppsLoadedAction } from '../../Actions/Profile';
 
 export const setCredentialsEpic = (action$, store, { profileRepository }) =>
-  action$.ofType(SET_CREDENTIALS)
-    .switchMap(q => {
-      profileRepository.storeCredentials(q.payload.credentials);
-      const availableApps = profileRepository.getStoredAppNamesCredentials();
-      return Observable.of(availableAppsLoadedAction(availableApps));
-    });
+  action$
+    .pipe(
+      ofType(SET_CREDENTIALS),
+      switchMap(q => {
+        profileRepository.storeCredentials(q.payload.credentials);
+        const availableApps = profileRepository.getStoredAppNamesCredentials();
+        return of(availableAppsLoadedAction(availableApps));
+      })
+    );
